This commit includes a major refactor of the `World` class and the addition of a `WorldFactory` class. Now, this factory returns a world depending on a selection screen shown as the game starts. I debated putting this in the start screen, but it seemed like more work and a more messy codebase. Now, there is a different world class for debug and default. With this separation, the world class can be extended to many unit tests (the next one being one for landmass collision avoidance). Feel free to add more as needed.
The `WorldFactory` class has been added to create different types of worlds. It includes methods for selecting a world, converting a world type to a string, and creating a world based on the world type.

Several changes were made to other files to accommodate these changes, including modifications to `GlobalInputHandler_PG.h`, `ShipEnums_PG.h`, `EnemyShipHandler_PG.cpp`, `EnemyShipMovementHandler_PG.cpp`, `IslandMenu_PG.cpp`, `StartMenu_PG.cpp`, and `main.cpp`.

New classes `DefaultDebugWorld` and `DefaultWorld` were added, inheriting from the `World` class and overriding the `createWorld` and `updateGameLoop` functions.

The project files `SFML Project 1.vcxproj` and `SFML Project 1.vcxproj.filters` were updated to include the new and modified files.

+++ PirateGame/Resource/EnemyShipHandler_PG.cpp
@@ -187,7 +187,7 @@ void EnemyShipHandler::interactWithNearbyShips(std::shared_ptr<ShipGroup> enemyS
 	if (enemyShipGroup->getIsInteracting()) return;
 
 	// Otherwise, roll a coin to see if the ship should be added to the group. 1 is a grouping, 2 is an attack, all other values are no interaction.
-	std::uniform_int_distribution<int> dist(0, interactionChance + enemyShipGroup->getEnemyShips().size()); // The more ships in the group, the less likely it is to interact (to prevent large groups from becoming too large)
+	std::uniform_int_distribution<int> dist(0, interactionChance + static_cast<int>(enemyShipGroup->getEnemyShips().size())); // The more ships in the group, the less likely it is to interact (to prevent large groups from becoming too large)
 	int interaction = dist(GlobalValues::getInstance().getRandomEngine());
 
 	// Shows if there is interaction. Possible framework for future attack indicator!
